% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipped.R
\name{d_coin_multiplicative}
\alias{d_coin_multiplicative}
\title{Compute probability of observations given an exponential decay model}
\usage{
d_coin_multiplicative(
  nheads,
  nflips,
  multiplier,
  log = FALSE,
  possibilities = get_possibilities(nheads, nflips),
  outside_bounds_is_NA = FALSE
)
}
\arguments{
\item{nheads}{Number of heads}

\item{nflips}{Total number of flips (heads and tails)}

\item{multiplier}{How much to multiply by each flip}

\item{log}{If TRUE return log transformed probabilities.}

\item{possibilities}{All possible sequences of flips that lead to the observed number of heads}

\item{outside_bounds_is_NA}{If TRUE, if any probability of heads is outside the bounds of probability, the function returns NA. Otherwise, it sets the value to the nearer bound.}
}
\value{
The likelihood of the data (or log likelihood if log=TRUE)
}
\description{
Compute probability of observations given an exponential decay model
}
