% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipped.R
\name{dcoin_exponential}
\alias{dcoin_exponential}
\title{Compute probability of observations given an exponential model
The idea is that the coin before handling has 100\% chance of heads, but each time it is picked up that probability will decrease (maybe it is bent by the statistician's mighty thumb). After halflife times handling it, the probability of heads is 50\%, and it keeps dropping from there.}
\usage{
dcoin_exponential(
  nheads,
  nflips,
  halflife,
  log = FALSE,
  possibilities = get_possibilities(nheads, nflips)
)
}
\arguments{
\item{nheads}{Number of heads}

\item{nflips}{Total number of flips (heads and tails)}

\item{halflife}{How many flips to get to 50\% heads}

\item{log}{If TRUE return log transformed probabilities.}

\item{possibilities}{All possible sequences of flips that lead to the observed number of heads}
}
\value{
The likelihood of the data (or log likelihood if log=TRUE)
}
\description{
Compute probability of observations given an exponential model
The idea is that the coin before handling has 100\% chance of heads, but each time it is picked up that probability will decrease (maybe it is bent by the statistician's mighty thumb). After halflife times handling it, the probability of heads is 50\%, and it keeps dropping from there.
}
