% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipped.R
\name{dcoin_from_probability}
\alias{dcoin_from_probability}
\title{Compute probability of observations given a vector of probability of heads}
\usage{
dcoin_from_probability(
  pheads,
  nheads,
  nflips,
  log = FALSE,
  possibilities = get_possibilities(nheads, nflips),
  diff_value = NULL
)
}
\arguments{
\item{pheads}{Vector with the probability of a heads on flip 1, 2, etc.}

\item{nheads}{Number of heads}

\item{nflips}{Total number of flips (heads and tails)}

\item{log}{If TRUE return log transformed probabilities.}

\item{possibilities}{All possible sequences of flips that lead to the observed number of heads}

\item{diff_value}{If not NULL, the final likelihood will be abs(likelihood - diff_value) for minimizing a function}
}
\value{
The likelihood of the data (or log likelihood if log=TRUE)
}
\description{
Compute probability of observations given a vector of probability of heads
}
