% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipped.R
\name{find_congruent_models}
\alias{find_congruent_models}
\title{Find congruent models to a simple binomial model
This will find the parameter values for other models that equal the likelihood for a simple binomial model. This may not be the MLE for these other models}
\usage{
find_congruent_models(
  nheads,
  nflips,
  slopes = c(0, 0.1, -0.05),
  stopval = 1e-04
)
}
\arguments{
\item{nheads}{Number of heads}

\item{nflips}{Total number of flips (heads and tails)}

\item{slopes}{Vector of slopes to use}

\item{stopval}{How large a difference in probability is considered close enough between the flat model and others}
}
\value{
A list containing the parameter estimates with likelihoods for each model and the probabilities for heads at each model
}
\description{
Find congruent models to a simple binomial model
This will find the parameter values for other models that equal the likelihood for a simple binomial model. This may not be the MLE for these other models
}
