% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipped.R
\name{get_possibilities}
\alias{get_possibilities}
\title{Exhaustively get all possible sets of outcomes that result in a specified number of heads out of a certain number of flips}
\usage{
get_possibilities(nheads, nflips)
}
\arguments{
\item{nheads}{Number of heads}

\item{nflips}{Total number of flips (heads and tails)}
}
\value{
data.frame with each potential trial as a row. 1=heads, 0=tails.
}
\description{
This grows very large with the number of flips. It will throw an error if you try too many flips.
}
