% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipped.R
\name{prob_heads_exponential}
\alias{prob_heads_exponential}
\title{Compute the probability of heads with each flip given an exponential model
The model assumes 100\% chance of heads before a coin is picked up and it drops exponentially each time the coin is handled.}
\usage{
prob_heads_exponential(nflips, halflife)
}
\arguments{
\item{nflips}{Total number of flips (heads and tails)}

\item{halflife}{How many flips to get to 50\% heads}
}
\value{
Vector of probability of heads for the first flip, second flip, etc.
}
\description{
Compute the probability of heads with each flip given an exponential model
The model assumes 100\% chance of heads before a coin is picked up and it drops exponentially each time the coin is handled.
}
