% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipped.R
\name{prob_heads_linear}
\alias{prob_heads_linear}
\title{Compute the probability of heads with each flip given a linear change model.}
\usage{
prob_heads_linear(nflips, preflip_prob = 0.5, slope = 0.1)
}
\arguments{
\item{nflips}{Total number of flips (heads and tails)}

\item{preflip_prob}{Probability of heads before the coin is handled}

\item{slope}{How much the probability changes each time the coin is flipped}
}
\value{
Vector of probability of heads for the first flip, second flip, etc.
}
\description{
Compute the probability of heads with each flip given a linear change model.
}
