% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipped.R
\name{prob_heads_multiplicative}
\alias{prob_heads_multiplicative}
\title{Compute the probability of heads with each flip given a multiplier model
The model assumes 50\% chance of heads before a coin is picked up and it changes as a percentage of the previous value each flip. i.e., the probability of heads is 101\% of the probability the previous flip with a multiplier of 1.01.}
\usage{
prob_heads_multiplicative(nflips, multiplier, outside_bounds_is_NA = FALSE)
}
\arguments{
\item{nflips}{Total number of flips (heads and tails)}

\item{multiplier}{Factor to multiply the previous probability by}

\item{outside_bounds_is_NA}{If TRUE, if any probability of heads is outside the bounds of probability, the function returns NA. Otherwise, it sets the value to the nearer bound.}
}
\value{
Vector of probability of heads for the first flip, second flip, etc.
}
\description{
Compute the probability of heads with each flip given a multiplier model
The model assumes 50\% chance of heads before a coin is picked up and it changes as a percentage of the previous value each flip. i.e., the probability of heads is 101\% of the probability the previous flip with a multiplier of 1.01.
}
