% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipped.R
\name{profile_exponential_decay_model}
\alias{profile_exponential_decay_model}
\title{Computes the likelihood for a range of values using an exponential coin model}
\usage{
profile_exponential_decay_model(
  nheads,
  nflips,
  param_range = c(0, nflips * 10),
  number_of_steps = 1000,
  log = FALSE
)
}
\arguments{
\item{nflips}{Total number of flips (heads and tails)}

\item{param_range}{Range of parameters to try}

\item{number_of_steps}{How many values of the parameter to try}
}
\value{
vector of likelihoods
}
\description{
Computes the likelihood for a range of values using an exponential coin model
}
\examples{
nheads <- 8
nflips <- 10
exp_results <- profile_exponential_decay_model(nheads, nflips)
plot(x=exp_results$preflip_prob, y=exp_results$likelihood, type="l")
best_param <- exp_results$halflife[which.max(exp_results$likelihood, na.rm=TRUE)]
print(best_param)
}
