% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipped.R
\name{profile_linear_model}
\alias{profile_linear_model}
\title{Computes the likelihood for a range of values using a linear coin model}
\usage{
profile_linear_model(
  nheads,
  nflips,
  param_range = c(0, 1),
  slope = 0.1,
  number_of_steps = 1000,
  log = FALSE,
  outside_bounds_is_NA = TRUE
)
}
\arguments{
\item{nflips}{Total number of flips (heads and tails)}

\item{param_range}{Range of parameters to try}

\item{slope}{How much the probability changes each time the coin is flipped}

\item{number_of_steps}{How many values of the parameter to try}
}
\value{
vector of likelihoods
}
\description{
Computes the likelihood for a range of values using a linear coin model
}
\examples{
nheads <- 8
nflips <- 10
linear_results <- profile_linear_model(nheads, nflips)
plot(x=linear_results$preflip_prob, y=linear_results$likelihood, type="l")
dbinom_proportions <- seq(from=0, to=1, length.out=1000)
lines(dbinom_proportions, dbinom(nheads, nflips, dbinom_proportions), col="red")
best_param <- linear_results$preflip_prob[which.max(linear_results$likelihood, na.rm=TRUE)]
print(best_param)
}
