% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipped.R
\name{try_many_vectors}
\alias{try_many_vectors}
\title{Compute probability of observations across many potential vectors
This will try (1/stepsize)^nflips possible vectors, computing the probability of the observation for each}
\usage{
try_many_vectors(
  nheads,
  nflips,
  number_samples = 1000,
  stopval = 1e-05,
  log = FALSE,
  possibilities = get_possibilities(nheads, nflips)
)
}
\arguments{
\item{nheads}{Number of heads}

\item{nflips}{Total number of flips (heads and tails)}

\item{number_samples}{How many vectors to sample}

\item{stopval}{How large a difference in probability is considered close enough between the flat model and others}

\item{log}{If TRUE return log transformed probabilities.}

\item{possibilities}{All possible sequences of flips that lead to the observed number of heads}
}
\value{
The likelihood of the data (or log likelihood if log=TRUE)
}
\description{
Compute probability of observations across many potential vectors
This will try (1/stepsize)^nflips possible vectors, computing the probability of the observation for each
}
